refactor: organize posapp modules
