+++ posawesome/public/js/posapp/components/navbar/index.js
@@ -1,0 +1,1 @@
+export { default as Navbar } from './Navbar.vue';
