+++ posawesome/public/js/posapp/components/offline-invoices/index.js
@@ -1,0 +1,1 @@
+export { default as OfflineInvoices } from './OfflineInvoices.vue';
