+++ posawesome/public/js/posapp/components/payments/index.js
@@ -1,0 +1,1 @@
+export { default as PaymentMethods } from './PaymentMethods.vue';
