+++ posawesome/public/js/posapp/components/pos/invoice/index.js
@@ -1,0 +1,3 @@
+export { default as Invoice } from './Invoice.vue';
+export { default as InvoiceSummary } from './InvoiceSummary.vue';
+export { default as ItemsTable } from './ItemsTable.vue';
