+++ posawesome/public/js/posapp/components/pos/orders/index.js
@@ -1,0 +1,1 @@
+export { default as SalesOrders } from './SalesOrders.vue';
