+++ posawesome/public/js/posapp/components/pos/returns/index.js
@@ -1,0 +1,1 @@
+export { default as Returns } from './Returns.vue';
