+++ posawesome/public/js/posapp/mixins/index.js
@@ -1,0 +1,6 @@
+export { default as responsiveMixin } from './responsive.js';
+export { default as invoiceComputed } from './invoiceComputed.js';
+export { default as invoiceItemMethods } from './invoiceItemMethods.js';
+export { default as invoiceOfferMethods } from './invoiceOfferMethods.js';
+export { default as invoiceShortcuts } from './invoiceShortcuts.js';
+export { default as invoiceWatchers } from './invoiceWatchers.js';
