+++ posawesome/public/js/posapp/mixins/invoiceShortcuts.js
